requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
decorator==5.1.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.29.0
jedi==0.19.1
matplotlib-inline==0.1.7
msgpack==1.1.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
redis==5.2.0
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.1.1
